+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [2020-07-05] Added rudimentary validation of request tokens by [@lflare].
 
 ### Changed
 - [2020-07-03] Updated README.md with relevant up-to-date information by [@lflare].